+++ NEWS.md
@@ -2,6 +2,7 @@
 
 # development version
 
+* Add Zhao 2014 ciprofloxacin ([doi:10.1128/AAC.03568-14](https://doi.org/10.1128/AAC.03568-14)) -- neonates and young infants less than three months of age receiving intravenous ciprofloxacin (ratifies new `CONMED_INOTROPE` covariate canonical).
 * Add Berges 2007 enoxaparin ([doi:10.1111/j.1365-2125.2007.02920.x](https://doi.org/10.1111/j.1365-2125.2007.02920.x)) -- elderly inpatients (>75 years) on prophylactic subcutaneous enoxaparin 4000 IU once daily for VTE prophylaxis (PROPHRE.75 study).
 * Add Padoin 1998 cephalexin ([doi:10.1128/aac.42.6.1463](https://doi.org/10.1128/aac.42.6.1463)) -- preclinical male Wistar rats with a cephalexin / quinapril oral coadministration DDI (ratifies new `CONMED_QPRL_ORAL` covariate canonical).
 * Add van Rongen 2018 metformin ([doi:10.1007/s40272-018-0293-1](https://doi.org/10.1007/s40272-018-0293-1)) -- overweight and obese Caucasian adolescents.